rename confidence->credible file
